hotspots: Modify draft hotspot to activate when first draft is created.
Usually, a hotspot is activated one after another
until it reaches the last hotspot.

In this commit, we modify the draft hotspot to
activate when the first time draft is created.
It will help new users to get familiar with the
draft system exactly when the first time they
create a draft.

To implement this,
In the backend, We update the hotspots event
to always send a draft hotspot until it is not seen
because it will help us not to rely on previous
hotspots to get seen first.

In the frontend, We avoid calling draft hotspot by
default instead, we trigger it just after the first
draft is created.